requirements.txt
PyYAML
Jinja2
snowflake-connector-python
colorama
snowflake-snowpark-python
snowflake-snowpark-python[pandas]
colorlog